feat: tests runner and initial test impl
Initial implementation of a tests runner; currently only makes sure that the TypeScript compiler successfully compiles build outputs. In the future, this is to be replaced by a functional tests system that runs tests on both the API and the extension. This just serves as a template for standardization of systems as of now.